+++ sdk/js/src/event/validator/zod.ts
@@ -1,0 +1,9 @@
+import { ZodSchema, z } from "zod";
+
+export function ZodValidator<Schema extends ZodSchema>(
+  schema: Schema
+): (input: z.input<Schema>) => z.output<Schema> {
+  return (input) => {
+    return schema.parse(input);
+  };
+}
